[SYCL] Enable LIT testing with CUDA BE
