+++ sycl/test/basic_tests/buffer/buffer_full_copy.cpp
@@ -6,9 +6,6 @@
 // RUN: %GPU_RUN_PLACEHOLDER %t2.out
 // RUN: %ACC_RUN_PLACEHOLDER %t2.out
 
-// TODO: cuda_piEnqueueMemBufferCopy not implemented
-// XFAIL: cuda
-
 //==------------- buffer_full_copy.cpp - SYCL buffer basic test ------------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
