+++ sycl/test/kernel_from_file/hw.cpp
@@ -6,9 +6,6 @@
 
 // TODO: InvalidTargetTriple: Expects spir-unknown-unknown or spir64-unknown-unknown. Actual target triple is x86_64-unknown-linux-gnu
 
-// XFAIL: cuda
-// Currently unsupported on cuda as this test specifically tests a SPV path.
-
 #include <CL/sycl.hpp>
 #include <iostream>
