+++ sycl/test/regression/static-buffer-dtor.cpp
@@ -16,7 +16,6 @@
 
 // TODO: terminate called after throwing an instance of 'cl::sycl::runtime_error'
 // TODO: what():  OpenCL API failed. OpenCL API returns: -999 (Unknown OpenCL error code) -999 (Unknown OpenCL error code)
-// XFAIL: cuda
 
 #include <CL/sycl.hpp>
