+++ sycl/test/usm/badmalloc.cpp
@@ -1,9 +1,9 @@
-// RUN: %clangxx -fsycl %s -o %t1.out
+// RUN: %clangxx -fsycl -fsycl-targets=%sycl_triple %s -o %t1.out
 // RUN: env SYCL_DEVICE_TYPE=HOST %t1.out
 // RUN: %CPU_RUN_PLACEHOLDER %t1.out
 // RUN: %GPU_RUN_PLACEHOLDER %t1.out
 
-// UNSUPPORTED: windows
+// UNSUPPORTED: windows,cuda
 
 //==----------------- badmalloc.cpp - Bad Mallocs test ---------------------==//
 //
