+++ sycl/test/usm/memset.cpp
@@ -1,6 +1,7 @@
-// RUN: %clangxx -fsycl %s -o %t1.out
+// RUN: %clangxx -fsycl -fsycl-targets=%sycl_triple  %s -o %t1.out
 // RUN: %CPU_RUN_PLACEHOLDER %t1.out
 // RUN: %GPU_RUN_PLACEHOLDER %t1.out
+// UNSUPPORTED: cuda
 
 //==---- memset.cpp - USM memset test --------------------------------------==//
 //
