+++ sycl/test/usm/mixed.cpp
@@ -2,6 +2,7 @@
 // RUN: env SYCL_DEVICE_TYPE=HOST %t1.out
 // RUN: %CPU_RUN_PLACEHOLDER %t1.out
 // RUN: %GPU_RUN_PLACEHOLDER %t1.out
+// UNSUPPORTED: cuda
 
 //==------------------- mixed.cpp - Mixed Memory test ---------------------==//
 //
