+++ sycl/test/usm/mixed2template.cpp
@@ -1,7 +1,8 @@
-// RUN: %clangxx -fsycl %s -o %t1.out
+// RUN: %clangxx -fsycl -fsycl-targets=%sycl_triple  %s -o %t1.out
 // RUN: env SYCL_DEVICE_TYPE=HOST %t1.out
 // RUN: %CPU_RUN_PLACEHOLDER %t1.out
 // RUN: %GPU_RUN_PLACEHOLDER %t1.out
+// UNSUPPORTED: cuda
 
 //==---------- mixed2template.cpp - Mixed Memory with Templatestest --------==//
 //
