+++ sycl/unittests/pi/cuda/CMakeLists.txt
@@ -5,6 +5,7 @@ add_sycl_unittest(PiCudaTests
   test_device.cpp
   test_kernels.cpp
   test_mem_obj.cpp
+  test_primary_context.cpp
   test_queue.cpp
   test_events.cpp
 )
